# server/openapi.yaml
openapi: 3.0.0
info:
  title: Fake StackOverflow REST API
  version: 1.0.0
  description: |
    API for managing questions, answers, and tags of a fake stack overflow application.

    The API is organized into three main routes:
    * Tags route: Managing tags and their statistics
    * Questions route: Managing questions and their operations
    * Answers route: Managing answers to questions
servers:
  - url: http://localhost:8000
    description: Development server for the Fake StackOverflow REST API.
...
components:
  responses:
...
    BadRequest:
      description: Bad Request - Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Internal Server Error"
  schemas:
    Tag:
      type: object
      properties:
        _id:
          type: string
          example: "679846196243c96505ba4007"
        name:
          type: string
          nullable: true
          example: "react"
        __v:
          type: integer
          example: 0
...
    Answer:
      type: object
      properties:
        _id:
          type: string
          example: "679d61405af1815cc9398a86"
        text:
          type: string
          example: You can solve your problem this way.
        ans_by:
          type: string
          example: user-name
        ans_date_time:
          type: string
          format: date-time
          example: "2025-02-02T00:28:19.406Z"
        __v:
          type: integer
          example: 0
        comments:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              user:
                type: string
              comment_date_time:
                type: string
                format: date-time
        upVotes:
          type: array
          items:
            type: string
        downVotes:
          type: array
          items:
            type: string
    Question:
      type: object
      properties:
        _id:
          type: string
          example: 679846196243c96505ba4007
        title:
          type: string
          example: Question title
        text:
          type: string
          example: This explains the question further
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        answers:
          type: array
          items:
            # oneOf is used to allow both answer IDs and answer objects answers in a question
            oneOf:
              - type: string # For answer IDs
                example: "679846196243c96505ba4007"
              - $ref: "#/components/schemas/Answer"
                example:
                  _id: "679d61405af1815cc9398a86"
                  text: "You can solve your problem this way."
                  ans_by: "user-name"
                  ans_date_time: "2025-02-02T00:28:19.406Z"
        asked_by:
          type: string
          example: user-name
        ask_date_time:
          type: string
          format: date-time
          example: "2025-02-02T00:28:19.406Z"
        views:
          type: integer
          default: 0
          example: 0
        __v:
          type: integer
          example: 0
    UserProfile:
      type: object
      properties:
        _id:
          type: string
          example: "645a1c08b302fca521f64a1b"
        username:
          type: string
          example: "john_doe"
        email:
          type: string
          example: "john@example.com"
        password:
          type: string
          example: "$2b$10$encryptedpassword"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
        comments:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
              text:
                type: string
              user:
                type: string
              comment_date_time:
                type: string
                format: date-time
...
    SignupRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          format: password
          example: "password123"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: "john_doe"
        password:
          type: string
          format: password
          example: "password123"
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "An error occurred"
paths:
...
  /user/signup:
    post:
      tags:
        - User
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User signed up successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User created successfully"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /user/login:
    post:
      tags:
        - User
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: User logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid username or password"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /user/profile:
    get:
      tags:
        - User
      summary: Get authenticated user's profile
      description: Requires a valid HTTP-only cookie (token)
      responses:
        "200":
          description: User profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized - Cookie is missing or invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /user/logout:
   post:
    tags:
      - User
    summary: Log out the current user
    responses:
      "200":
        description: User logged out successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  example: "User logged out successfully"
      "500":
        $ref: "#/components/responses/InternalServerError"